pyproject.toml
[project]
name = "deepclaude"
version = "0.1.0"
description = "This is a project for combining DeepSeek and Claude into one api call. Full unleash the power of AI."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.11",
    "colorlog>=6.9.0",
    "fastapi>=0.115.8",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "tiktoken>=0.8.0",
    "uvicorn>=0.34.0",
]
